+++ tests/fixtures/option-construct-byvalue-return-spec/expected.c
@@ -9687,7 +9687,7 @@ static void __tur_module_def_init(void) {
     SCHEMA_unAP_1501 = INT64_C(14);
     SCHEMA_unFIELD_1502 = INT64_C(15);
     SCHEMA_unAP_unFAT_1503 = INT64_C(16);
-    PI_1646 = 3.14159265358979;
+    PI_1646 = 3.141592653589793;
 }
 
 static void __tur_fatbox_init(void) {